+++ tests/integration/test_spans.py
@@ -45,7 +45,6 @@ def test_span_extraction(
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:indexed-spans-extraction",
     ]
     project_config["config"]["transactionMetrics"] = {
@@ -244,8 +243,8 @@ def test_span_extraction(
 @pytest.mark.parametrize(
     "sample_rate,expected_spans,expected_metrics",
     [
-        (None, 2, 7),
-        (1.0, 2, 7),
+        (None, 2, 3),
+        (1.0, 2, 3),
         (0.0, 0, 0),
     ],
 )
@@ -266,7 +265,6 @@ def test_span_extraction_with_sampling(
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:indexed-spans-extraction",
     ]
     project_config["config"]["transactionMetrics"] = {
@@ -311,7 +309,6 @@ def test_duplicate_performance_score(mini_sentry, relay):
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:indexed-spans-extraction",
     ]
     project_config["config"]["transactionMetrics"] = {
@@ -683,7 +680,6 @@ def test_span_ingestion(
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
         "organizations:standalone-span-ingestion",
-        "projects:span-metrics-extraction",
         "projects:relay-otel-endpoint",
     ]
     project_config["config"]["transactionMetrics"] = {
@@ -1167,182 +1163,6 @@ def test_span_ingestion(
             "value": 4.0,
             "received_at": time_after(now_timestamp),
         },
-        {
-            "name": "d:spans/duration@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "retention_days": 90,
-            "tags": {
-                "file_extension": "js",
-                "span.category": "resource",
-                "span.description": "https://example.com/*/blah.js",
-                "span.domain": "example.com",
-                "span.group": "8a97a9e43588e2bd",
-                "span.op": "resource.script",
-            },
-            "timestamp": expected_timestamp + 1,
-            "type": "d",
-            "value": [1500.0, 1500.0],
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "name": "d:spans/duration@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "retention_days": 90,
-            "tags": {
-                "span.category": "db",
-                "span.op": "default",
-            },
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0, 500.0],
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "name": "d:spans/duration@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "retention_days": 90,
-            "tags": {
-                "span.op": "default",
-            },
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0, 500.0],
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "name": "d:spans/duration@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "retention_days": 90,
-            "tags": {"span.op": "default"},
-            "timestamp": expected_timestamp + 1,
-            "type": "d",
-            "value": [1500.0, 1500.0],
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "name": "d:spans/duration_light@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "received_at": time_after(now_timestamp),
-            "retention_days": 90,
-            "tags": {
-                "file_extension": "js",
-                "span.category": "resource",
-                "span.description": "https://example.com/*/blah.js",
-                "span.domain": "example.com",
-                "span.group": "8a97a9e43588e2bd",
-                "span.op": "resource.script",
-            },
-            "timestamp": expected_timestamp + 1,
-            "type": "d",
-            "value": [1500.0, 1500.0],
-        },
-        {
-            "name": "d:spans/duration_light@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "received_at": time_after(now_timestamp),
-            "retention_days": 90,
-            "tags": {"span.category": "db", "span.op": "default"},
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0, 500.0],
-        },
-        {
-            "org_id": 1,
-            "project_id": 42,
-            "name": "d:spans/exclusive_time@millisecond",
-            "type": "d",
-            "value": [161.0, 345.0],
-            "timestamp": expected_timestamp + 1,
-            "tags": {
-                "file_extension": "js",
-                "span.category": "resource",
-                "span.description": "https://example.com/*/blah.js",
-                "span.domain": "example.com",
-                "span.group": "8a97a9e43588e2bd",
-                "span.op": "resource.script",
-            },
-            "retention_days": 90,
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "org_id": 1,
-            "project_id": 42,
-            "name": "d:spans/exclusive_time@millisecond",
-            "retention_days": 90,
-            "tags": {"span.category": "db", "span.op": "default"},
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0, 500.0],
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "name": "d:spans/exclusive_time@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "retention_days": 90,
-            "tags": {"span.op": "default"},
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0, 500.0],
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "org_id": 1,
-            "project_id": 42,
-            "name": "d:spans/exclusive_time@millisecond",
-            "retention_days": 90,
-            "tags": {"span.op": "default"},
-            "timestamp": expected_timestamp + 1,
-            "type": "d",
-            "value": [345.0, 345.0],
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "org_id": 1,
-            "project_id": 42,
-            "name": "d:spans/exclusive_time_light@millisecond",
-            "type": "d",
-            "value": [161.0, 345.0],
-            "timestamp": expected_timestamp + 1,
-            "tags": {
-                "file_extension": "js",
-                "span.category": "resource",
-                "span.description": "https://example.com/*/blah.js",
-                "span.domain": "example.com",
-                "span.group": "8a97a9e43588e2bd",
-                "span.op": "resource.script",
-            },
-            "retention_days": 90,
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "name": "d:spans/exclusive_time_light@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "retention_days": 90,
-            "tags": {"span.category": "db", "span.op": "default"},
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0, 500.0],
-            "received_at": time_after(now_timestamp),
-        },
-        {
-            "name": "d:spans/webvital.score.total@ratio",
-            "org_id": 1,
-            "project_id": 42,
-            "retention_days": 90,
-            "tags": {"span.op": "resource.script"},
-            "timestamp": expected_timestamp + 1,
-            "type": "d",
-            "value": [0.12121616],
-            "received_at": time_after(now_timestamp),
-        },
     ]
 
     span_metrics = [m for m in metrics if ":spans/" in m["name"]]
@@ -1350,15 +1170,6 @@ def test_span_ingestion(
     assert len(span_metrics) == len(expected_span_metrics)
     for actual, expected in zip(span_metrics, expected_span_metrics):
         assert actual == expected
-
-    # Regardless of whether transactions are extracted, score.total is only converted to a transaction metric once:
-    score_total_metrics = [
-        m
-        for m in metrics
-        if m["name"] == "d:transactions/measurements.score.total@ratio"
-    ]
-    assert len(score_total_metrics) == 1, score_total_metrics
-    assert len(score_total_metrics[0]["value"]) == 1
 
     metrics_consumer.assert_empty()
 
@@ -1385,7 +1196,6 @@ def test_standalone_span_ingestion_metric_extraction(
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
         "organizations:standalone-span-ingestion",
-        "projects:span-metrics-extraction",
         # "projects:relay-otel-endpoint",
     ]
 
@@ -1463,74 +1273,6 @@ def test_standalone_span_ingestion_metric_extraction(
             "timestamp": expected_timestamp,
             "type": "c",
             "value": 1.0,
-        },
-        {
-            "name": "d:spans/duration@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "received_at": expected_received,
-            "retention_days": 90,
-            "tags": {
-                "span.action": "SELECT",
-                "span.category": "db",
-                "span.description": "SELECT from",
-                "span.group": "e7ef86adbb98803e",
-                "span.op": "db",
-            },
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0],
-        },
-        {
-            "name": "d:spans/duration_light@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "received_at": expected_received,
-            "retention_days": 90,
-            "tags": {
-                "span.action": "SELECT",
-                "span.op": "db",
-                "span.category": "db",
-                "span.description": "SELECT from",
-                "span.group": "e7ef86adbb98803e",
-            },
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0],
-        },
-        {
-            "name": "d:spans/exclusive_time@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "received_at": expected_received,
-            "retention_days": 90,
-            "tags": {
-                "span.action": "SELECT",
-                "span.category": "db",
-                "span.description": "SELECT from",
-                "span.group": "e7ef86adbb98803e",
-                "span.op": "db",
-            },
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0],
-        },
-        {
-            "name": "d:spans/exclusive_time_light@millisecond",
-            "org_id": 1,
-            "project_id": 42,
-            "received_at": expected_received,
-            "retention_days": 90,
-            "tags": {
-                "span.action": "SELECT",
-                "span.op": "db",
-                "span.category": "db",
-                "span.description": "SELECT from",
-                "span.group": "e7ef86adbb98803e",
-            },
-            "timestamp": expected_timestamp,
-            "type": "d",
-            "value": [500.0],
         },
     ]
 
@@ -1760,7 +1502,6 @@ def test_span_ingestion_with_performance_scores(
     project_config["config"]["features"] = [
         "organizations:performance-calculate-score-relay",
         "organizations:standalone-span-ingestion",
-        "projects:span-metrics-extraction",
     ]
     project_config["config"]["txNameRules"] = [
         {
@@ -1992,7 +1733,6 @@ def test_rate_limit_indexed_consistent(
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:standalone-span-ingestion",
     ]
     project_config["config"]["quotas"] = [
@@ -2051,7 +1791,6 @@ def test_rate_limit_consistent_extracted(
         "version": TRANSACTION_EXTRACT_MIN_SUPPORTED_VERSION
     }
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:indexed-spans-extraction",
     ]
     project_config["config"]["quotas"] = [
@@ -2099,7 +1838,7 @@ def test_rate_limit_consistent_extracted(
     # one for the transaction, one for the contained span
     assert len(spans) == 2
     # A limit only for span_indexed does not affect extracted metrics
-    metrics = metrics_consumer.get_metrics(n=11)
+    metrics = metrics_consumer.get_metrics(n=7)
     span_count = sum(
         [m[0]["value"] for m in metrics if m[0]["name"] == "c:spans/usage@none"]
     )
@@ -2140,7 +1879,6 @@ def test_rate_limit_spans_in_envelope(
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:standalone-span-ingestion",
     ]
     project_config["config"]["quotas"] = [
@@ -2198,7 +1936,6 @@ def test_rate_limit_is_consistent_between_transaction_and_spans(
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:standalone-span-ingestion",
         "organizations:indexed-spans-extraction",
     ]
@@ -2336,7 +2073,6 @@ def test_span_extraction_with_tags(
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:indexed-spans-extraction",
     ]
 
@@ -2441,6 +2177,9 @@ def test_dynamic_sampling(
     }
 
     sampling_config = mini_sentry.add_basic_project_config(43)
+    sampling_config["config"]["features"] = [
+        "organizations:standalone-span-ingestion",
+    ]
     sampling_public_key = sampling_config["publicKeys"][0]["publicKey"]
     sampling_config["config"]["txNameRules"] = [
         {
@@ -2509,103 +2248,15 @@ def test_dynamic_sampling(
         outcomes_consumer.assert_empty()
     else:
         outcomes = outcomes_consumer.get_outcomes(timeout=10, n=1)
-        assert summarize_outcomes(outcomes) == {(16, 1): 6}  # Span, Filtered
-        assert {o["reason"] for o in outcomes} == {"Sampled:3000"}
+        assert summarize_outcomes(outcomes) == {
+            (16, 1): 6,  # SpanIndexed, Filtered
+        }
+        assert {o["reason"] for o in outcomes} == {
+            "Sampled:3000",
+        }
 
     spans_consumer.assert_empty()
     outcomes_consumer.assert_empty()
-
-
-@pytest.mark.parametrize("ingest_in_eap", [True, False])
-def test_ingest_in_eap_for_organization(
-    mini_sentry,
-    relay_with_processing,
-    spans_consumer,
-    ingest_in_eap,
-):
-    spans_consumer = spans_consumer()
-
-    relay = relay_with_processing(options=TEST_CONFIG)
-    project_id = 42
-    project_config = mini_sentry.add_full_project_config(project_id)
-    project_config["config"]["features"] = [
-        "organizations:indexed-spans-extraction",
-    ]
-
-    if ingest_in_eap:
-        project_config["config"]["features"] += ["organizations:ingest-spans-in-eap"]
-
-    event = make_transaction({"event_id": "cbf6960622e14a45abc1f03b2055b186"})
-    end = datetime.now(timezone.utc) - timedelta(seconds=1)
-    duration = timedelta(milliseconds=500)
-    start = end - duration
-    event["spans"] = [
-        {
-            "description": "GET /api/0/organizations/?member=1",
-            "op": "http",
-            "origin": "manual",
-            "parent_span_id": "968cff94913ebb07",
-            "span_id": "bbbbbbbbbbbbbbbb",
-            "start_timestamp": start.isoformat(),
-            "status": "success",
-            "timestamp": end.isoformat(),
-            "trace_id": "ff62a8b040f340bda5d830223def1d81",
-        },
-    ]
-
-    relay.send_event(project_id, event)
-
-    if ingest_in_eap:
-        spans_consumer.get_span()
-        spans_consumer.get_span()
-
-    spans_consumer.assert_empty()
-
-
-@pytest.mark.parametrize("ingest_in_eap", [True, False])
-def test_ingest_in_eap_for_project(
-    mini_sentry,
-    relay_with_processing,
-    spans_consumer,
-    ingest_in_eap,
-):
-    spans_consumer = spans_consumer()
-
-    relay = relay_with_processing(options=TEST_CONFIG)
-    project_id = 42
-    project_config = mini_sentry.add_full_project_config(project_id)
-    project_config["config"]["features"] = [
-        "organizations:indexed-spans-extraction",
-    ]
-
-    if ingest_in_eap:
-        project_config["config"]["features"] += ["projects:ingest-spans-in-eap"]
-
-    event = make_transaction({"event_id": "cbf6960622e14a45abc1f03b2055b186"})
-    end = datetime.now(timezone.utc) - timedelta(seconds=1)
-    duration = timedelta(milliseconds=500)
-    start = end - duration
-    event["spans"] = [
-        {
-            "description": "GET /api/0/organizations/?member=1",
-            "op": "http",
-            "origin": "manual",
-            "parent_span_id": "968cff94913ebb07",
-            "span_id": "bbbbbbbbbbbbbbbb",
-            "start_timestamp": start.isoformat(),
-            "status": "success",
-            "timestamp": end.isoformat(),
-            "trace_id": "ff62a8b040f340bda5d830223def1d81",
-        },
-    ]
-
-    relay.send_event(project_id, event)
-
-    if ingest_in_eap:
-        spans_consumer.get_span()
-        spans_consumer.get_span()
-
-    spans_consumer.assert_empty()
 
 
 @pytest.mark.parametrize("scrub_ip_addresses", [False, True])
@@ -2622,7 +2273,6 @@ def test_scrubs_ip_addresses(
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
-        "projects:span-metrics-extraction",
         "organizations:indexed-spans-extraction",
     ]
     project_config["config"].setdefault("datascrubbingSettings", {})[