+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
+
 ## Unreleased
 
 **Internal**:
 
 - Emit outcomes for skipped large attachments on playstation crashes. ([#4862](https://github.com/getsentry/relay/pull/4862))
+- Disable span metrics. ([#4931](https://github.com/getsentry/relay/pull/4931))
 
 ## 25.7.0